Move datatype utility functions to own file.
