+++ src/theory/datatypes/datatypes_rewriter.cpp
@@ -17,6 +17,8 @@
 #include "theory/datatypes/datatypes_rewriter.h"
 
 #include "expr/node_algorithm.h"
+#include "theory/datatypes/theory_datatypes_utils.h"
+#include "options/datatypes_options.h"
 
 using namespace CVC4;
 using namespace CVC4::kind;
@@ -55,7 +57,7 @@ RewriteResponse DatatypesRewriter::postRewrite(TNode in)
         }
       }
       TNode constructor = in[0].getOperator();
-      size_t constructorIndex = indexOf(constructor);
+      size_t constructorIndex = utils::indexOf(constructor);
       const Datatype& dt = Datatype::datatypeOf(constructor.toExpr());
       const DatatypeConstructor& c = dt[constructorIndex];
       unsigned weight = c.getWeight();
@@ -118,7 +120,7 @@ RewriteResponse DatatypesRewriter::postRewrite(TNode in)
     {
       Trace("dt-sygus-util") << "Rewrite " << in << " by unfolding...\n";
       const Datatype& dt = ev.getType().getDatatype();
-      unsigned i = indexOf(ev.getOperator());
+      unsigned i = utils::indexOf(ev.getOperator());
       Node op = Node::fromExpr(dt[i].getSygusOp());
       // if it is the "any constant" constructor, return its argument
       if (op.getAttribute(SygusAnyConstAttribute()))
@@ -141,9 +143,9 @@ RewriteResponse DatatypesRewriter::postRewrite(TNode in)
         cc.insert(cc.end(), args.begin(), args.end());
         children.push_back(nm->mkNode(DT_SYGUS_EVAL, cc));
       }
-      Node ret = mkSygusTerm(dt, i, children);
+      Node ret = utils::mkSygusTerm(dt, i, children);
       // apply the appropriate substitution
-      ret = applySygusArgs(dt, op, ret, args);
+      ret = utils::applySygusArgs(dt, op, ret, args);
       Trace("dt-sygus-util") << "...got " << ret << "\n";
       return RewriteResponse(REWRITE_AGAIN_FULL, ret);
     }
@@ -215,7 +217,7 @@ RewriteResponse DatatypesRewriter::postRewrite(TNode in)
       }
       if (!cons.isNull())
       {
-        cases.push_back(mkTester(h, cindex, dt));
+        cases.push_back(utils::mkTester(h, cindex, dt));
       }
       else
       {
@@ -246,7 +248,7 @@ RewriteResponse DatatypesRewriter::postRewrite(TNode in)
       return RewriteResponse(REWRITE_DONE, nm->mkConst(true));
     }
     std::vector<Node> rew;
-    if (checkClash(in[0], in[1], rew))
+    if (utils::checkClash(in[0], in[1], rew))
     {
       Trace("datatypes-rewrite")
           << "Rewrite clashing equality " << in << " to false" << std::endl;
@@ -269,154 +271,6 @@ RewriteResponse DatatypesRewriter::postRewrite(TNode in)
   }
 
   return RewriteResponse(REWRITE_DONE, in);
-}
-
-Node DatatypesRewriter::applySygusArgs(const Datatype& dt,
-                                       Node op,
-                                       Node n,
-                                       const std::vector<Node>& args)
-{
-  if (n.getKind() == BOUND_VARIABLE)
-  {
-    Assert(n.hasAttribute(SygusVarNumAttribute()));
-    int vn = n.getAttribute(SygusVarNumAttribute());
-    Assert(Node::fromExpr(dt.getSygusVarList())[vn] == n);
-    return args[vn];
-  }
-  // n is an application of operator op.
-  // We must compute the free variables in op to determine if there are
-  // any substitutions we need to make to n.
-  TNode val;
-  if (!op.hasAttribute(SygusVarFreeAttribute()))
-  {
-    std::unordered_set<Node, NodeHashFunction> fvs;
-    if (expr::getFreeVariables(op, fvs))
-    {
-      if (fvs.size() == 1)
-      {
-        for (const Node& v : fvs)
-        {
-          val = v;
-        }
-      }
-      else
-      {
-        val = op;
-      }
-    }
-    Trace("dt-sygus-fv") << "Free var in " << op << " : " << val << std::endl;
-    op.setAttribute(SygusVarFreeAttribute(), val);
-  }
-  else
-  {
-    val = op.getAttribute(SygusVarFreeAttribute());
-  }
-  if (val.isNull())
-  {
-    return n;
-  }
-  if (val.getKind() == BOUND_VARIABLE)
-  {
-    // single substitution case
-    int vn = val.getAttribute(SygusVarNumAttribute());
-    TNode sub = args[vn];
-    return n.substitute(val, sub);
-  }
-  // do the full substitution
-  std::vector<Node> vars;
-  Node bvl = Node::fromExpr(dt.getSygusVarList());
-  for (unsigned i = 0, nvars = bvl.getNumChildren(); i < nvars; i++)
-  {
-    vars.push_back(bvl[i]);
-  }
-  return n.substitute(vars.begin(), vars.end(), args.begin(), args.end());
-}
-
-Kind DatatypesRewriter::getOperatorKindForSygusBuiltin(Node op)
-{
-  Assert(op.getKind() != BUILTIN);
-  if (op.getKind() == LAMBDA)
-  {
-    return APPLY_UF;
-  }
-  TypeNode tn = op.getType();
-  if (tn.isConstructor())
-  {
-    return APPLY_CONSTRUCTOR;
-  }
-  else if (tn.isSelector())
-  {
-    return APPLY_SELECTOR;
-  }
-  else if (tn.isTester())
-  {
-    return APPLY_TESTER;
-  }
-  else if (tn.isFunction())
-  {
-    return APPLY_UF;
-  }
-  return UNDEFINED_KIND;
-}
-
-Node DatatypesRewriter::mkSygusTerm(const Datatype& dt,
-                                    unsigned i,
-                                    const std::vector<Node>& children)
-{
-  Trace("dt-sygus-util") << "Make sygus term " << dt.getName() << "[" << i
-                         << "] with children: " << children << std::endl;
-  Assert(i < dt.getNumConstructors());
-  Assert(dt.isSygus());
-  Assert(!dt[i].getSygusOp().isNull());
-  std::vector<Node> schildren;
-  Node op = Node::fromExpr(dt[i].getSygusOp());
-  Trace("dt-sygus-util") << "Operator is " << op << std::endl;
-  if (children.empty())
-  {
-    // no children, return immediately
-    Trace("dt-sygus-util") << "...return direct op" << std::endl;
-    return op;
-  }
-  // if it is the any constant, we simply return the child
-  if (op.getAttribute(SygusAnyConstAttribute()))
-  {
-    Assert(children.size() == 1);
-    return children[0];
-  }
-  if (op.getKind() != BUILTIN)
-  {
-    schildren.push_back(op);
-  }
-  schildren.insert(schildren.end(), children.begin(), children.end());
-  Node ret;
-  if (op.getKind() == BUILTIN)
-  {
-    ret = NodeManager::currentNM()->mkNode(op, schildren);
-    Trace("dt-sygus-util") << "...return (builtin) " << ret << std::endl;
-    return ret;
-  }
-  Kind ok = NodeManager::operatorToKind(op);
-  Trace("dt-sygus-util") << "operator kind is " << ok << std::endl;
-  if (ok != UNDEFINED_KIND)
-  {
-    // If it is an APPLY_UF operator, we should have at least an operator and
-    // a child.
-    Assert(ok != APPLY_UF || schildren.size() != 1);
-    ret = NodeManager::currentNM()->mkNode(ok, schildren);
-    Trace("dt-sygus-util") << "...return (op) " << ret << std::endl;
-    return ret;
-  }
-  Kind tok = getOperatorKindForSygusBuiltin(op);
-  if (schildren.size() == 1 && tok == kind::UNDEFINED_KIND)
-  {
-    ret = schildren[0];
-  }
-  else
-  {
-    ret = NodeManager::currentNM()->mkNode(tok, schildren);
-  }
-  Trace("dt-sygus-util") << "...return " << ret << std::endl;
-  return ret;
 }
 
 RewriteResponse DatatypesRewriter::preRewrite(TNode in)
@@ -443,7 +297,7 @@ RewriteResponse DatatypesRewriter::preRewrite(TNode in)
         Node op = in.getOperator();
         // get the constructor object
         const DatatypeConstructor& dtc =
-            Datatype::datatypeOf(op.toExpr())[indexOf(op)];
+            Datatype::datatypeOf(op.toExpr())[utils::indexOf(op)];
         // create ascribed constructor type
         Node tc = NodeManager::currentNM()->mkConst(
             AscriptionType(dtc.getSpecializedConstructorType(t)));
@@ -496,7 +350,7 @@ RewriteResponse DatatypesRewriter::rewriteSelector(TNode in)
     TypeNode argType = in[0].getType();
     Expr selector = in.getOperator().toExpr();
     TNode constructor = in[0].getOperator();
-    size_t constructorIndex = indexOf(constructor);
+    size_t constructorIndex = utils::indexOf(constructor);
     const Datatype& dt = Datatype::datatypeOf(selector);
     const DatatypeConstructor& c = dt[constructorIndex];
     Trace("datatypes-rewrite-debug") << "Rewriting collapsable selector : "
@@ -600,7 +454,7 @@ RewriteResponse DatatypesRewriter::rewriteTester(TNode in)
 {
   if (in[0].getKind() == kind::APPLY_CONSTRUCTOR)
   {
-    bool result = indexOf(in.getOperator()) == indexOf(in[0].getOperator());
+    bool result = utils::indexOf(in.getOperator()) == utils::indexOf(in[0].getOperator());
     Trace("datatypes-rewrite") << "DatatypesRewriter::postRewrite: "
                                << "Rewrite trivial tester " << in << " "
                                << result << std::endl;
@@ -620,189 +474,6 @@ RewriteResponse DatatypesRewriter::rewriteTester(TNode in)
   }
   // could try dt.getNumConstructors()==2 && indexOf(in.getOperator())==1 ?
   return RewriteResponse(REWRITE_DONE, in);
-}
-
-bool DatatypesRewriter::checkClash(Node n1, Node n2, std::vector<Node>& rew)
-{
-  Trace("datatypes-rewrite-debug") << "Check clash : " << n1 << " " << n2
-                                   << std::endl;
-  if (n1.getKind() == kind::APPLY_CONSTRUCTOR
-      && n2.getKind() == kind::APPLY_CONSTRUCTOR)
-  {
-    if (n1.getOperator() != n2.getOperator())
-    {
-      Trace("datatypes-rewrite-debug") << "Clash operators : " << n1 << " "
-                                       << n2 << " " << n1.getOperator() << " "
-                                       << n2.getOperator() << std::endl;
-      return true;
-    }
-    Assert(n1.getNumChildren() == n2.getNumChildren());
-    for (unsigned i = 0, size = n1.getNumChildren(); i < size; i++)
-    {
-      if (checkClash(n1[i], n2[i], rew))
-      {
-        return true;
-      }
-    }
-  }
-  else if (n1 != n2)
-  {
-    if (n1.isConst() && n2.isConst())
-    {
-      Trace("datatypes-rewrite-debug") << "Clash constants : " << n1 << " "
-                                       << n2 << std::endl;
-      return true;
-    }
-    else
-    {
-      Node eq = NodeManager::currentNM()->mkNode(kind::EQUAL, n1, n2);
-      rew.push_back(eq);
-    }
-  }
-  return false;
-}
-/** get instantiate cons */
-Node DatatypesRewriter::getInstCons(Node n, const Datatype& dt, int index)
-{
-  Assert(index >= 0 && index < (int)dt.getNumConstructors());
-  std::vector<Node> children;
-  NodeManager* nm = NodeManager::currentNM();
-  children.push_back(Node::fromExpr(dt[index].getConstructor()));
-  Type t = n.getType().toType();
-  for (unsigned i = 0, nargs = dt[index].getNumArgs(); i < nargs; i++)
-  {
-    Node nc = nm->mkNode(kind::APPLY_SELECTOR_TOTAL,
-                         Node::fromExpr(dt[index].getSelectorInternal(t, i)),
-                         n);
-    children.push_back(nc);
-  }
-  Node n_ic = nm->mkNode(kind::APPLY_CONSTRUCTOR, children);
-  if (dt.isParametric())
-  {
-    TypeNode tn = TypeNode::fromType(t);
-    // add type ascription for ambiguous constructor types
-    if (!n_ic.getType().isComparableTo(tn))
-    {
-      Debug("datatypes-parametric") << "DtInstantiate: ambiguous type for "
-                                    << n_ic << ", ascribe to " << n.getType()
-                                    << std::endl;
-      Debug("datatypes-parametric") << "Constructor is " << dt[index]
-                                    << std::endl;
-      Type tspec =
-          dt[index].getSpecializedConstructorType(n.getType().toType());
-      Debug("datatypes-parametric") << "Type specification is " << tspec
-                                    << std::endl;
-      children[0] = nm->mkNode(kind::APPLY_TYPE_ASCRIPTION,
-                               nm->mkConst(AscriptionType(tspec)),
-                               children[0]);
-      n_ic = nm->mkNode(kind::APPLY_CONSTRUCTOR, children);
-      Assert(n_ic.getType() == tn);
-    }
-  }
-  Assert(isInstCons(n, n_ic, dt) == index);
-  // n_ic = Rewriter::rewrite( n_ic );
-  return n_ic;
-}
-
-int DatatypesRewriter::isInstCons(Node t, Node n, const Datatype& dt)
-{
-  if (n.getKind() == kind::APPLY_CONSTRUCTOR)
-  {
-    int index = indexOf(n.getOperator());
-    const DatatypeConstructor& c = dt[index];
-    Type nt = n.getType().toType();
-    for (unsigned i = 0, size = n.getNumChildren(); i < size; i++)
-    {
-      if (n[i].getKind() != kind::APPLY_SELECTOR_TOTAL
-          || n[i].getOperator() != Node::fromExpr(c.getSelectorInternal(nt, i))
-          || n[i][0] != t)
-      {
-        return -1;
-      }
-    }
-    return index;
-  }
-  return -1;
-}
-
-int DatatypesRewriter::isTester(Node n, Node& a)
-{
-  if (n.getKind() == kind::APPLY_TESTER)
-  {
-    a = n[0];
-    return indexOf(n.getOperator());
-  }
-  return -1;
-}
-
-int DatatypesRewriter::isTester(Node n)
-{
-  if (n.getKind() == kind::APPLY_TESTER)
-  {
-    return indexOf(n.getOperator().toExpr());
-  }
-  return -1;
-}
-
-struct DtIndexAttributeId
-{
-};
-typedef expr::Attribute<DtIndexAttributeId, uint64_t> DtIndexAttribute;
-
-unsigned DatatypesRewriter::indexOf(Node n)
-{
-  if (!n.hasAttribute(DtIndexAttribute()))
-  {
-    Assert(n.getType().isConstructor() || n.getType().isTester()
-           || n.getType().isSelector());
-    unsigned index = Datatype::indexOfInternal(n.toExpr());
-    n.setAttribute(DtIndexAttribute(), index);
-    return index;
-  }
-  return n.getAttribute(DtIndexAttribute());
-}
-
-Node DatatypesRewriter::mkTester(Node n, int i, const Datatype& dt)
-{
-  return NodeManager::currentNM()->mkNode(
-      kind::APPLY_TESTER, Node::fromExpr(dt[i].getTester()), n);
-}
-
-Node DatatypesRewriter::mkSplit(Node n, const Datatype& dt)
-{
-  std::vector<Node> splits;
-  for (unsigned i = 0, ncons = dt.getNumConstructors(); i < ncons; i++)
-  {
-    Node test = mkTester(n, i, dt);
-    splits.push_back(test);
-  }
-  NodeManager* nm = NodeManager::currentNM();
-  return splits.size() == 1 ? splits[0] : nm->mkNode(kind::OR, splits);
-}
-
-bool DatatypesRewriter::isNullaryApplyConstructor(Node n)
-{
-  Assert(n.getKind() == kind::APPLY_CONSTRUCTOR);
-  for (unsigned i = 0; i < n.getNumChildren(); i++)
-  {
-    if (n[i].getType().isDatatype())
-    {
-      return false;
-    }
-  }
-  return true;
-}
-
-bool DatatypesRewriter::isNullaryConstructor(const DatatypeConstructor& c)
-{
-  for (unsigned j = 0, nargs = c.getNumArgs(); j < nargs; j++)
-  {
-    if (c[j].getType().getRangeType().isDatatype())
-    {
-      return false;
-    }
-  }
-  return true;
 }
 
 Node DatatypesRewriter::normalizeCodatatypeConstant(Node n)
